@@ -12,20 +12,20 @@
     "@sveltejs/kit": "next",
     "prettier": "~2.2.1",
     "prettier-plugin-svelte": "^2.2.0",
-    "svelte": "^3.29.0",
-    "svelte-preprocess": "^4.0.0",
-    "tslib": "^2.0.0",
-    "typescript": "^4.0.0",
-    "vite": "^2.1.0",
-    "vite-plugin-windicss": "^0.15.7"
+    "svelte": "^3.38.2",
+    "svelte-preprocess": "^4.7.3",
+    "tslib": "^2.2.0",
+    "typescript": "^4.2.4",
+    "vite": "^2.2.4",
+    "vite-plugin-windicss": "^0.15.10"
   },
   "type": "module",
   "dependencies": {
-    "@fontsource/fira-mono": "^4.2.2",
+    "@fontsource/fira-mono": "^4.3.0",
     "@lukeed/uuid": "^2.0.0",
     "@types/mime": "^2.0.3",
     "axios": "^0.21.1",
-    "cheerio": "^1.0.0-rc.6",
+    "cheerio": "^1.0.0-rc.9",
     "cookie": "^0.4.1",
     "fast-xml-parser": "^3.19.0",
     "fs-extra": "^9.1.0",
