@@ -1,5 +1,5 @@
 import { v4 as uuid } from '@lukeed/uuid';
-import type { GetContext, Handle } from '@sveltejs/kit';
+import type { Handle } from '@sveltejs/kit';
 import cookie from 'cookie';
 import { ReqZoneMiddleware } from './util/reqZone';
 import { promises as fs } from 'fs';
@@ -13,33 +13,23 @@ type Context = {
 };
 /** 默认的 handle 处理器 */
 export type defaultHandle = Handle<Context>;
-export const getContext: GetContext<Context> = (request) => {
-	const cookies = cookie.parse(request.headers.cookie || '');
 
-	return {
-		is_new: !cookies.userid,
-		userid: cookies.userid || uuid()
-	};
-};
-const rawHandel = async ({ request, render }) => {
+const rawHandel: Handle = async ({ request, render }) => {
+	const cookies = cookie.parse(request.headers.cookie || '');
+	request.locals.userid = cookies.userid || uuid();
+
 	// TODO https://github.com/sveltejs/kit/issues/1046
-	const response = await render({
-		...request,
-		method: (request.query.get('_method') || request.method).toUpperCase()
-	});
-	const { is_new, userid } = request.context;
-	if (is_new) {
-		// if this is the first time the user has visited this app,
-		// set a cookie so that we recognise them when they return
-		return {
-			...response,
-			headers: {
-				...response.headers,
-				'set-cookie': `userid=${userid}; Path=/; HttpOnly`
-			}
-		};
+	if (request.query.has('_method')) {
+		request.method = request.query.get('_method').toUpperCase();
 	}
 
+	const response = await render(request);
+
+	if (!cookies.userid) {
+		// if this is the first time the user has visited this app,
+		// set a cookie so that we recognise them when they return
+		response.headers['set-cookie'] = `userid=${request.locals.userid}; Path=/; HttpOnly`;
+	}
 	return response;
 };
 /** 静态文件服务 */
